Resolving the „Dokcer Host Panic“ Issue #204
Issue 204: Docker experiences a panic response when encountering unauthorized or invalid Client Tokens in the users.db file. In response, it attempts to reconnect every 10 seconds. However, upon initial startup, the system checks for a 401 response. If a 401 response is detected, the reconnection process for this websocket is terminated, effectively resolving the panic issue experienced by the Docker host.